Revert "Reapply "🔥 댓글 추가 기능까지 완료""
This reverts commit a1d108b7581dc7da7604a85749dd07f2c4fe96a9.
